// sanity.types.ts
export type Product = {
  _id: string
  _type: 'product'
  _createdAt: string
  _updatedAt: string
  _rev: string
  name?: string
  slug?: Slug
  images?: Array<{
    asset?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
    }
    media?: unknown
    hotspot?: SanityImageHotspot
    crop?: SanityImageCrop
    _type: 'image'
    _key: string
  }>
  intro?: string
  description?: string
  price?: number
  discount?: number
  categories?: Array<{
    _ref: string
    _type: 'reference'
    _weak?: boolean
    _key: string
    [internalGroqTypeReferenceTo]?: 'category'
  }>
  stock?: number
  status?: 'new' | 'hot' | 'sale'
  variant?: 'tshirt' | 'jacket' | 'pants' | 'hoodie' | 'short' | 'others'
}
// ...
export type Slug = {
  _type: 'slug'
  current?: string
  source?: string
}
// ...
export type SanityImageCrop = {
  _type: 'sanity.imageCrop'
  top?: number
  bottom?: number
  left?: number
  right?: number
}

export type SanityImageHotspot = {
  _type: 'sanity.imageHotspot'
  x?: number
  y?: number
  height?: number
  width?: number
}
// ...
export declare const internalGroqTypeReferenceTo: unique symbol